+++ tests/core/base/test_config_base.py
@@ -168,6 +168,7 @@ class TestBasePipelineConfig:
             "current_date",
             "framework_version",
             "py_version",
+            "image_uri",
             "source_dir",
             "enable_caching",
             "use_secure_pypi",